SPI_Tx-Rx_with_Arduino_code

+++ STM32Fxx_drivers/drivers/Src/stm32f407xx_spi_drivers.c
@@ -161,7 +161,7 @@ void SPI_ReceiveData(SPI_RegDef_t *pSPIx, uint8_t *pRxBuffer, uint32_t Len)
 	while(Len > 0)
 		{
 			//1. wait until RXNE is set
-			while(SPI_GetFlagStatus(pSPIx,SPI_RXNE_FLAG)  == (uint8_t)FLAG_RESET );
+			while(SPI_GetFlagStatus(pSPIx,SPI_RXNE_FLAG)  == FLAG_RESET );
 
 			//2. check the DFF bit in CR1
 			if( (pSPIx->CR1 & ( 1 << SPI_CR1_DFF) ) )
